cleanup and move to scripts.js
