+++ aemedge/blocks/fragment/fragment.js
@@ -5,41 +5,12 @@
  */
 
 import {
-  decorateMain,
+  loadFragment,
 } from '../../scripts/scripts.js';
 
-import {
-  loadBlocks,
-} from '../../scripts/aem.js';
-
-/**
- * Loads a fragment.
- * @param {string} path The path to the fragment
- * @returns {HTMLElement} The root element of the fragment
- */
-export async function loadFragment(path) {
-  if (path && path.startsWith('/')) {
-    const resp = await fetch(`${path}.plain.html`);
-    if (resp.ok) {
-      const main = document.createElement('main');
-      main.innerHTML = await resp.text();
-
-      // reset base path for media to fragment base
-      const resetAttributeBase = (tag, attr) => {
-        main.querySelectorAll(`${tag}[${attr}^="./media_"]`).forEach((elem) => {
-          elem[attr] = new URL(elem.getAttribute(attr), new URL(path, window.location)).href;
-        });
-      };
-      resetAttributeBase('img', 'src');
-      resetAttributeBase('source', 'srcset');
-
-      await decorateMain(main, false);
-      await loadBlocks(main);
-      return main;
-    }
-  }
-  return null;
-}
+export {
+  loadFragment,
+} from '../../scripts/scripts.js';
 
 export default async function decorate(block) {
   const link = block.querySelector('a');
